Improve download_source and simplify Dockerfiles
download_source script is rewritten to be self-encapsulated with
a much clearer usage. This greatly simplifies the Dockerfile.
All Dockerfile has to do now is to copy the script, run it once,
install stuff, run it again, and finally remove it--no additional
custom logic. At this level, the dockerfiles tool can even
automatically add these statements into the Dockerfile.